% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_zeroinfl.R
\name{BIC.zeroinfl}
\alias{BIC.zeroinfl}
\title{Bayesian information criterion implementation for \code{zeroinfl} objects}
\usage{
\method{BIC}{zeroinfl}(object)
}
\description{
Bayesian information criterion implementation for \code{zeroinfl} objects
}
