% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latextable.R
\name{extract_coeff}
\alias{extract_coeff}
\title{Wrapper to extract coefficients from a list of models}
\usage{
extract_coeff(idx, model_list, modeltype)
}
\arguments{
\item{idx}{Index}

\item{model_list}{List of models}

\item{modeltype}{Character vectors indicating whether
we should use selection or outcome model. Ignored if
model is not zeroinfl}
}
\value{
List with coefficients ready to be merged
 to produce a latex table
}
\description{
Wrapper to extract coefficients from a list of models
}
