% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastglm_nb.R
\name{fastglm.nb}
\alias{fastglm.nb}
\title{Fit a negative binomial generalized linear model}
\usage{
\method{fastglm}{nb}(formula, data, weights, subset, na.action,
  start = NULL, etastart, mustart, control = glm.control(...),
  method = "glm.fit", model = TRUE, x = FALSE, y = TRUE,
  contrasts = NULL, ..., init.theta = NULL, link = log)
}
\arguments{
\item{formula}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{data}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{weights}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{subset}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{na.action}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{start}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{etastart}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{mustart}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{control}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{method}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{model}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{x}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{y}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{contrasts}{
  arguments for the \code{\link{glm}()} function.
  Note that these exclude \code{family} and \code{offset}
  (but \code{\link{offset}()} can be used).
}

\item{init.theta}{
  Optional initial value for the theta parameter.  If omitted a moment
  estimator after an initial fit using a Poisson GLM is used.
}

\item{link}{
  The link function.  Currently must be one of \code{log}, \code{sqrt}
  or \code{identity}.
}
}
\description{
A modification of the function \link[MASS]{glm.nb} to
 reduce time and memory consumption
}
\seealso{
\link{fastglm.nb} ; \link[MASS]{theta.ml} ; \link[MASS]{glm.nb}
}
