% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastzeroinfl.R
\name{fastzeroinfl}
\alias{fastzeroinfl}
\title{Fast implementation for zero-inflated Count Data Regression}
\usage{
fastzeroinfl(formula, data, subset, na.action, weights, offset,
  dist = c("poisson", "negbin", "geometric"), link = c("logit",
  "probit", "cloglog", "cauchit", "log"),
  control = pscl::zeroinfl.control(...), model = TRUE, y = TRUE,
  x = FALSE, ...)
}
\description{
Fast implementation for zero-inflated Count Data Regression
}
