% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_package.R
\name{gravity-exports}
\alias{gravity-exports}
\title{gravity exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of gravity.}
\description{
gravity exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of gravity.
}
\keyword{internal}
