% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info}
\alias{info}
\title{Information iteration value}
\usage{
info(n, th, mu, y)
}
\arguments{
\item{n}{Number of observations}

\item{th}{\eqn{\theta} parameter}

\item{mu}{Predicted values}

\item{y}{Observed values}
}
\description{
Information iteration value
}
