% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_distance.R
\name{log_distance}
\alias{log_distance}
\title{Apply logarithm to distance column}
\usage{
log_distance(data, distance)
}
\arguments{
\item{data}{(Type: data.frame) the dataset to be used.}

\item{distance}{The distance column}
}
\value{
The function returns the summary of the estimated gravity model as an
\code{\link[stats]{lm}}-object.
}
\description{
\code{log_distance} creates a new \code{dist_log}
column in log scale by taking the original distance column in the data
}
\examples{
log_distance(gravity_zeros, "distw")
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[lmtest]{coeftest}},
\code{\link[sandwich]{vcovHC}}
}
