% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score}
\alias{score}
\title{Score for negative binomial maximum-likelihood estimates}
\usage{
score(n, th, mu, y)
}
\arguments{
\item{n}{Number of observations}

\item{th}{\eqn{\theta} parameter}

\item{mu}{Predicted values}

\item{y}{Observed values}
}
\value{
Score value
}
\description{
Score for negative binomial maximum-likelihood estimates
}
