% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speed_theta_ml}
\alias{speed_theta_ml}
\title{Maximum likelihood estimation for negative binomial models}
\usage{
speed_theta_ml(y, mu, limit = 10L, eps = 1e-06, trace = TRUE)
}
\arguments{
\item{y}{
    Vector of observed values from the Negative Binomial.
  }

\item{mu}{
    Estimated mean vector.
  }

\item{limit}{
    Limit on the number of iterations.
  }

\item{eps}{
    Tolerance to determine convergence.
  }

\item{trace}{
    logical: should iteration progress be printed?
  }
}
\value{
A maximum-likelihood estimator for \eqn{\theta}
}
\description{
Simplified C++ function for \link[MASS]{theta.ml}
}
\references{
<Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer>
}
\seealso{
\link{fastglm.nb} ; \link[MASS]{theta.ml} ; \link[MASS]{glm.nb}
}
