% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimobjects.R
\name{strip}
\alias{strip}
\alias{strip.glm}
\alias{strip.summary.glm}
\alias{strip.zeroinfl}
\alias{strip.summary.zeroinfl}
\title{Strip regression objects to reduce memory needs}
\usage{
strip(object, ...)

\method{strip}{glm}(object)

\method{strip}{summary.glm}(object)

\method{strip}{zeroinfl}(object)

\method{strip}{summary.zeroinfl}(object)
}
\arguments{
\item{object}{A regression object or summary.
Accepted classes are \link[stats]{glm} or
\link[pscl]{zeroinfl}}
}
\value{
Initial object stripped from
 heavy elements. Initial class is returned
 with a new one indicating the object has
 been stripped.
}
\description{
Remove elements that are more related
 to prediction than model to reduce
 memory needs to store a glm output
}
