% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateLatex.R
\docType{data}
\name{templateLatex}
\alias{templateLatex}
\title{A model for \code{latex} table}
\format{An object of class \code{character} of length 51.}
\usage{
templateLatex
}
\description{
This is a template for a
 zero-inflated negative binomial
 model
}
\keyword{datasets}
