% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimobjects.R
\name{trim_big_object}
\alias{trim_big_object}
\title{Trim a big object to reduce memory need}
\usage{
trim_big_object(obj)
}
\arguments{
\item{obj}{An object whose class is 
accepted by \link{strip} method}
}
\value{
A list with following elements
\itemize{
  \item object - \link{strip} object
  \item summary - \link{strip} summary object. For
   \link[pscl]{zeroinfl} objects, contain elements
   \code{obj_trim}, \code{se_count}, \code{se_selection}
}
}
\description{
This function returns as little information
 from obj needed
}
