% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvw.R
\name{bvw}
\alias{bvw}
\title{Bonus vetus OLS (BVW)}
\usage{
bvw(dependent_variable, regressors, incomes, codes, vce_robust = TRUE, data,
  ...)
}
\arguments{
\item{dependent_variable}{name (type: character) of the dependent variable in the dataset 
\code{data} (i.e trade flows). 

This dependent variable is divided by the 
product of unilateral incomes (named \code{inc_o} and \code{inc_d}, e.g. 
GDPs or GNPs of the countries of interest) and logged afterwards.

The transformed variable is then used as the dependent variable in the 
estimation.}

\item{regressors}{name (type: character) of the distance variable in the dataset 
\code{data} containing a measure of distance between all pairs of bilateral
partners and bilateral variables that should be taken as the independent variables 
in the estimation. 

The distance is logged automatically when the function is executed.

If an independent variable is a dummy variable, it should be of type numeric (0/1) in the dataset.

If an independent variable is defined as a ratio, it should be logged.

Unilateral metric variables such as GDPs should be inserted via the argument \code{incomes}. 

As country specific effects are subdued due to demeaning, no further unilateral variables 
equivalent to \code{c(gdp origin, gdp destination)} can be added.

Write this argument as \code{c(distance, other variables)}.}

\item{incomes}{variable name (type: character) of the income of the country of 
origin (i.e \code{gdp_o}) and destination (i.e \code{gdp_d}) in the dataset \code{data}. 

The dependent variable \code{dependent_variable} is divided by the product of the incomes.

Write this argument as \code{c(origin, destination)}.}

\item{codes}{variable name (type: character) of the code of the country 
(i.e ISO-3 code) of origin (i.e \code{iso_o}) and destination (i.e \code{iso_d}) in the dataset \code{data}. 

The variables are grouped by using \code{code_o} and \code{code_d} to obtain estimates.

Write this argument as \code{c(origin, destination)}.}

\item{vce_robust}{robust (type: logical) determines whether a robust 
variance-covariance matrix should be used. By default is set to \code{TRUE}.

If \code{vce_robust = TRUE} the estimation results are consistent with the 
Stata code provided at \href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
when choosing robust estimation.}

\item{data}{name of the dataset to be used (type: character).

As an example, to estimate gravity equations you need a square dataset including bilateral 
flows defined by the argument \code{dependent_variable}, ISO-codes of type character 
(i.e. \code{iso_o} for the country of origin and \code{iso_d} for the 
destination country), a distance measure defined by the argument \code{distance} 
and other potential influences (i.e. contiguity or common currency) given as a vector in 
\code{regressors} are required.

All dummy variables should be of type numeric (0/1). Some of our functions remove observations with 
null/missing trade flows or distances as those variables are converted to log scale before obtaining estimates. 

The user should perform some data cleaning beforehand to remove observations that contain entries that 
can distort estimates.

As, to our knowledge at the moment, there is no explicit literature covering the estimation of a gravity 
equation by \code{bvu} using panel data, cross-sectional data should be used.}

\item{...}{additional arguments to be passed to \code{bvw}.}
}
\value{
The function returns the summary of the estimated gravity model as an 
\code{\link[stats]{lm}}-object.
}
\description{
\code{bvw} estimates gravity models via Bonus 
vetus OLS with GDP-weights.
}
\details{
Bonus vetus OLS is an estimation method for gravity models 
developed by \insertCite{Baier2009,Baier2010;textual}{gravity} using GDP-weights to center a
Taylor-series (see the references for more information). 

To execute the function a square gravity dataset with all pairs of 
countries, ISO-codes for the country of origin and destination, a measure of 
distance between the bilateral partners as well as all 
information that should be considered as dependent an independent 
variables is needed. 

Make sure the ISO-codes are of type "character".

Missing bilateral flows as well as incomplete rows should be 
excluded from the dataset.

Furthermore, flows equal to zero should be excluded as the gravity equation 
is estimated in its additive form.

The \code{bvw} function considers Multilateral Resistance terms and allows to 
conduct comparative statics. Country specific effects are subdued due 
to demeaning. Hence, unilateral variables apart from \code{inc_o}
and \code{inc_d} cannot be included in the estimation.

\code{bvw} is designed to be consistent with the Stata code provided at 
\href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
when choosing robust estimation.

As, to our knowledge at the moment, there is no explicit literature covering 
the estimation of a gravity equation by \code{bvw} using panel data, 
we do not recommend to apply this method in this case.
}
\examples{
\dontrun{
data(gravity_no_zeros)

bvw(dependent_variable = "flow", regressors = c("distw", "rta"), 
incomes = c("gdp_o", "gdp_d"), codes = c("iso_o", "iso_d"),
vce_robust = TRUE, data = gravity_no_zeros)

bvw(dependent_variable = "flow", regressors = c("distw", "rta", "comcur", "contig"), 
incomes = c("gdp_o", "gdp_d"), codes = c("iso_o", "iso_d"),
vce_robust = TRUE, data = gravity_no_zeros)
}

\dontshow{
# examples for CRAN checks:
# executable in < 5 sec together with the examples above
# not shown to users

data(gravity_no_zeros)
# choose exemplarily 10 biggest countries for check data
countries_chosen <- names(sort(table(gravity_no_zeros$iso_o), decreasing = TRUE)[1:10])
grav_small <- gravity_no_zeros[gravity_no_zeros$iso_o \%in\% countries_chosen,]
bvw(dependent_variable = "flow", regressors = c("distw", "rta"), 
incomes = c("gdp_o", "gdp_d"), codes = c("iso_o", "iso_d"),
vce_robust = TRUE, data = grav_small)
}

}
\references{
For estimating gravity equations via Bonus Vetus OLS see

\insertRef{Baier2009}{gravity}

\insertRef{Baier2010}{gravity}

For more information on gravity models, theoretical foundations and
estimation methods in general see 

\insertRef{Anderson1979}{gravity}

\insertRef{Anderson2001}{gravity}

\insertRef{Anderson2010}{gravity}

\insertRef{Head2010}{gravity}

\insertRef{Head2014}{gravity}

\insertRef{Santos2006}{gravity}

and the citations therein.

See \href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook} for gravity datasets and Stata code for estimating gravity models.
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[lmtest]{coeftest}}, 
\code{\link[sandwich]{vcovHC}}
}
