% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvw.R
\name{bvw}
\alias{bvw}
\title{Bonus vetus OLS (BVW)}
\usage{
bvw(dependent_variable, distance, additional_regressors = NULL,
  income_origin, income_destination, code_origin, code_destination,
  robust = FALSE, data, ...)
}
\arguments{
\item{dependent_variable}{(Type: character) name of the dependent variable in the dataset
\code{data} (e.g. trade flows).

This dependent variable is divided by the
product of unilateral incomes (e.g.
GDPs or GNPs of the countries of interest, named \code{inc_o} and \code{inc_d} in the example datasets)
and logged afterwards.

The transformed variable is then used as the dependent variable in the
estimation.}

\item{distance}{(Type: character) name of the distance variable in the dataset \code{data} containing a measure of
distance between all pairs of bilateral partners and bilateral variables that should
be taken as the independent variables in the estimation.

The distance is logged automatically when the function is executed.}

\item{additional_regressors}{(Type: character) names of the additional regressors to include in the model (e.g. a dummy
variable to indicate contiguity).

Unilateral metric variables such as GDPs should be inserted via the arguments \code{income_origin} and \code{income_origin}.

As country specific effects are subdued due to demeaning, no further unilateral variables apart from incomes can be added.

Write this argument as \code{c(contiguity, common currency, ...)}.}

\item{income_origin}{(Type: character) variable name of the income of the country of
origin (e.g. \code{inc_o}) in the dataset \code{data}. The dependent variable \code{dependent_variable} is
divided by the product of the incomes.}

\item{income_destination}{(Type: character) variable name of the income of the country of
destination (e.g. \code{inc_d}) in the dataset \code{data}. The dependent variable \code{dependent_variable} is
divided by the product of the incomes.}

\item{code_origin}{(Type: character) variable name of the code of the country
of origin (e.g. ISO-3 codes from the variables \code{iso_o} in the
example datasets). The variables are grouped by using \code{iso_o} and \code{iso_d} to obtain estimates.}

\item{code_destination}{(Type: character) variable name of the code of the country
of destination (e.g. ISO-3 codes from the variables \code{iso_d}) in the
example datasets). The variables are grouped by using \code{iso_o} and \code{iso_d} to obtain estimates.}

\item{robust}{(Type: logical) determines whether robust
fitting should be used. By default is set to \code{FALSE}.}

\item{data}{(Type: character) name of the dataset to be used.

To estimate gravity equations you need a square dataset including bilateral
flows defined by the argument \code{dependent_variable}, ISO codes or similar of type character
(e.g. \code{iso_o} for the country of origin and \code{iso_d} for the
destination country), a distance measure defined by the argument \code{distance}
and other potential influences (e.g. contiguity and common currency) given as a vector in
\code{regressors} are required.

All dummy variables should be of type numeric (0/1).

Make sure the origin and destination codes are of type "character".

If an independent variable is defined as a ratio, it should be logged.

The user should perform some data cleaning beforehand to remove observations that contain entries that
can distort estimates.

The function will remove zero flows and distances.}

\item{...}{Additional arguments to be passed to \code{bvw}.}
}
\value{
The function returns the summary of the estimated gravity model as an
\code{\link[stats]{lm}}-object.
}
\description{
\code{bvw} estimates gravity models via Bonus
vetus OLS with GDP-weights.
}
\details{
Bonus vetus OLS is an estimation method for gravity models
developed by \insertCite{Baier2009,Baier2010;textual}{gravity} using GDP-weights to center a
Taylor-series (see the references for more information).

The \code{bvw} function considers Multilateral Resistance terms and allows to
conduct comparative statics. Country specific effects are subdued due
to demeaning. Hence, unilateral variables apart from \code{inc_o}
and \code{inc_d} cannot be included in the estimation.

\code{bvw} is designed to be consistent with the Stata code provided at
\href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
when choosing robust estimation.

As, to our knowledge at the moment, there is no explicit literature covering
the estimation of a gravity equation by \code{bvw} using panel data,
we do not recommend to apply this method in this case.
}
\examples{
# Example for CRAN checks:
# Executable in < 5 sec
library(dplyr)
data("gravity_no_zeros")

# Choose 5 countries for testing
countries_chosen <- c("AUS", "CHN", "GBR", "BRA", "CAN")
grav_small <- filter(gravity_no_zeros, iso_o \%in\% countries_chosen)

fit <- bvw(
  dependent_variable = "flow",
  distance = "distw",
  additional_regressors = c("rta", "comcur", "contig"),
  income_origin = "gdp_o",
  income_destination = "gdp_d",
  code_origin = "iso_o",
  code_destination = "iso_d",
  robust = FALSE,
  data = grav_small
)

}
\references{
For estimating gravity equations via Bonus Vetus OLS see

\insertRef{Baier2009}{gravity}

\insertRef{Baier2010}{gravity}

For more information on gravity models, theoretical foundations and
estimation methods in general see

\insertRef{Anderson1979}{gravity}

\insertRef{Anderson2001}{gravity}

\insertRef{Anderson2010}{gravity}

\insertRef{Baier2009}{gravity}

\insertRef{Baier2010}{gravity}

\insertRef{Head2010}{gravity}

\insertRef{Head2014}{gravity}

\insertRef{Santos2006}{gravity}

and the citations therein.

See \href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
for gravity datasets and Stata code for estimating gravity models.
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[lmtest]{coeftest}},
\code{\link[sandwich]{vcovHC}}
}
