% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm}
\alias{ddm}
\title{Double Demeaning (DDM)}
\usage{
ddm(dependent_variable, distance, additional_regressors = NULL, code_origin,
  code_destination, robust = TRUE, data, ...)
}
\arguments{
\item{dependent_variable}{(Type: character) name of the dependent variable in the dataset
\code{data} (e.g. trade flows).

This dependent variable is divided by the
product of unilateral incomes (e.g.
GDPs or GNPs of the countries of interest, named \code{inc_o} and \code{inc_d} in the example datasets)
and logged afterwards.

The transformed variable is then used as the dependent variable in the
estimation.}

\item{distance}{(Type: character) name of the distance variable in the dataset \code{data} containing a measure of
distance between all pairs of bilateral partners and bilateral variables that should
be taken as the independent variables in the estimation.

The distance is logged automatically when the function is executed.}

\item{additional_regressors}{(Type: character) names of the additional regressors to include in the model (e.g. a dummy 
variable to indicate contiguity).

Unilateral metric variables such as GDPs should be inserted via the arguments \code{income_origin} and \code{income_origin}.

As country specific effects are subdued due to demeaning, no further unilateral variables apart from incomes can be added.

Write this argument as \code{c(contiguity, common currency, ...)}.}

\item{code_origin}{(Type: character) variable name of the code of the country
of origin (e.g. ISO-3 codes from the variables \code{iso_o} in the
example datasets). The variables are grouped by using \code{iso_o} and \code{iso_d} to obtain estimates.}

\item{code_destination}{(Type: character) variable name of the code of the country
of destination (e.g. ISO-3 codes from the variables \code{iso_d}) in the
example datasets). The variables are grouped by using \code{iso_o} and \code{iso_d} to obtain estimates.}

\item{robust}{(Type: logical) determines whether robust
fitting should be used. By default is set to \code{TRUE}.}

\item{data}{(Type: character) name of the dataset to be used.

To estimate gravity equations you need a square dataset including bilateral
flows defined by the argument \code{dependent_variable}, ISO codes or similar of type character
(e.g. \code{iso_o} for the country of origin and \code{iso_d} for the
destination country), a distance measure defined by the argument \code{distance}
and other potential influences (e.g. contiguity and common currency) given as a vector in
\code{regressors} are required.

All dummy variables should be of type numeric (0/1).

Make sure the ISO codes are of type "character".

If an independent variable is defined as a ratio, it should be logged.

The user should perform some data cleaning beforehand to remove observations that contain entries that
can distort estimates.

The function will remove zero flows and distances.}

\item{...}{additional arguments to be passed to \code{ddm}.}
}
\value{
The function returns the summary of the estimated gravity model as an
\code{\link[stats]{lm}}-object.
}
\description{
\code{ddm} estimates gravity models via double demeaning the
left hand side and right hand side of the gravity equation.
}
\details{
\code{ddm} is an estimation method for gravity models presented
in \insertCite{Head2014;textual}{gravity}.

Country specific effects are subdued due double demeaning. Hence, unilateral income
proxies such as GDP cannot be considered as exogenous variables.

Unilateral effect drop out due to double demeaning and therefore cannot be estimated.

\code{ddm} is designed to be consistent with the Stata code provided at
\href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
when choosing robust estimation.

As, to our knowledge at the moment, there is no explicit literature covering
the estimation of a gravity equation by \code{ddm} using panel data,
we do not recommend to apply this method in this case.
}
\examples{
\dontrun{
data(gravity_no_zeros)

ddm(dependent_variable = "flow", 
distance = "distw", additional_regressors = "rta",
code_origin = "iso_o", code_destination = "iso_d",
robust = TRUE, data = gravity_no_zeros)

ddm(dependent_variable = "flow", 
distance = "distw", additional_regressors = c("distw", "rta", "comcur", "contig"),
code_origin = "iso_o", code_destination = "iso_d",
robust = TRUE, data = gravity_no_zeros)
}

\dontshow{
# examples for CRAN checks:
# executable in < 5 sec together with the examples above
# not shown to users

data(gravity_no_zeros)
# choose exemplarily 10 biggest countries for check data
countries_chosen <- names(sort(table(gravity_no_zeros$iso_o), decreasing = TRUE)[1:10])
grav_small <- gravity_no_zeros[gravity_no_zeros$iso_o \%in\% countries_chosen,]
ddm(dependent_variable = "flow", distance = "distw", additional_regressors = "rta",
code_origin = "iso_o", code_destination = "iso_d",
robust = TRUE, data = grav_small)
}

}
\references{
For more information on Double Demeaning as well as information on gravity
models, theoretical foundations and estimation methods in general see

\insertRef{Head2014}{gravity}

as well as

\insertRef{Anderson1979}{gravity}

\insertRef{Anderson2001}{gravity}

\insertRef{Anderson2010}{gravity}

\insertRef{Baier2009}{gravity}

\insertRef{Baier2010}{gravity}

\insertRef{Head2010}{gravity}

\insertRef{Santos2006}{gravity}

and the citations therein.

See \href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook} 
for gravity datasets and Stata code for estimating gravity models.
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[lmtest]{coeftest}},
\code{\link[sandwich]{vcovHC}}
}
