% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects.R
\name{fixed_effects}
\alias{fixed_effects}
\title{Fixed Effects}
\usage{
fixed_effects(dependent_variable, distance, additional_regressors = NULL,
  code_origin, code_destination, robust = FALSE, data, ...)
}
\arguments{
\item{dependent_variable}{(Type: character) name of the dependent variable. This variable is logged and then used as 
the dependent variable in the estimation.}

\item{distance}{(Type: character) name of the distance variable that should be taken as the key independent variable 
in the estimation. The distance is logged automatically when the function is executed.}

\item{additional_regressors}{(Type: character) names of the additional regressors to include in the model (e.g. a dummy
variable to indicate contiguity). Unilateral metric variables such as GDP should be inserted via the arguments 
\code{income_origin} and \code{income_destination}.

Write this argument as \code{c(contiguity, common currency, ...)}. By default this is set to \code{NULL}.}

\item{code_origin}{(Type: character) country of origin variable (e.g. ISO-3 country codes). The variables are grouped 
using this parameter.}

\item{code_destination}{(Type: character) country of destination variable (e.g. country ISO-3 codes). The variables are 
grouped using this parameter.}

\item{robust}{(Type: logical) whether robust fitting should be used. By default this is set to \code{FALSE}.}

\item{data}{(Type: data.frame) the dataset to be used.}

\item{...}{Additional arguments to be passed to \code{bvw}.}
}
\value{
The function returns the summary of the estimated gravity model as an
\code{\link[stats]{lm}}-object.
}
\description{
\code{fixed_effects} estimates gravity models via
OLS and fixed effects for the countries of origin and destination.
}
\details{
To account for MR terms, \insertCite{Feenstra2002;textual}{gravity} proposes to use
importer and exporter fixed effects. Due to the use of these effects, all
unilateral influences such as GDPs can no longer be estimated.

A disadvantage of the use of \code{fixed_effects} is that, when applied to
panel data, the number of country-year or country-pair fixed effects can be
too high for estimation. In addition, no comparative statistics are
possible with \code{fixed_effects} as the Multilateral Resistance terms are not estimated
explicitly. Nevertheless, \insertCite{Head2014;textual}{gravity} highlight the importance of
the use of fixed effects.

By including country specific fixed effects, all monadic effects
are captured, including Multilateral Resistance terms.
Therefore, no other unilateral variables such as GDP can be
included as independent variables in the estimation.

\code{fixed_effects} estimation can be used for both, cross-sectional as well as
panel data.

Nonetheless, the function is designed to be consistent with the
Stata code for cross-sectional data provided at the website
\href{https://sites.google.com/site/hiegravity/}{Gravity Equations: Workhorse, Toolkit, and Cookbook}
when choosing robust estimation.

The function \code{fixed_effects} was therefore tested for
cross-sectional data. Its up to the user to ensure that the functions can be applied
to panel data.

Depending on the panel dataset and the variables -
specifically the type of fixed effects -
included in the model, it may easily occur that the model is not computable.

Also, note that by including bilateral fixed effects such as country-pair
effects, the coefficients of time-invariant observables such as distance
can no longer be estimated.

Depending on the specific model, the code of the
respective function may has to be changed in order to exclude the distance
variable from the estimation.

At the very least, the user should take special
care with respect to the meaning of the estimated coefficients and variances
as well as the decision about which effects to include in the estimation.

When using panel data, the parameter and variance estimation of the models
may have to be changed accordingly.

For a comprehensive overview of gravity models for panel data
see \insertCite{Egger2003;textual}{gravity}, \insertCite{Gomez-Herrera2013;textual}{gravity} and
\insertCite{Head2010;textual}{gravity} as well as the references therein.
}
\examples{
# Example for CRAN checks:
# Executable in < 5 sec
library(dplyr)
data("gravity_no_zeros")

# Choose 5 countries for testing
countries_chosen <- c("AUS", "CHN", "GBR", "BRA", "CAN")
grav_small <- filter(gravity_no_zeros, iso_o \%in\% countries_chosen)

fit <- fixed_effects(
  dependent_variable = "flow",
  distance = "distw",
  additional_regressors = c("rta", "comcur", "contig"),
  code_origin = "iso_o",
  code_destination = "iso_d",
  robust = FALSE,
  data = grav_small
)

}
\references{
\insertRef{Egger2003}{gravity}

\insertRef{Feenstra2002}{gravity}

\insertRef{Gomez-Herrera2013}{gravity}

\insertRef{Head2010}{gravity}

\insertRef{Head2014}{gravity}
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[lmtest]{coeftest}},
\code{\link[sandwich]{vcovHC}}
}
