% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity-package.R
\name{gravity-exports}
\alias{gravity-exports}
\alias{mutate}
\alias{\%>\%}
\alias{select}
\alias{mutate}
\alias{group_by}
\alias{ungroup}
\alias{row_number}
\alias{left_join}
\alias{gather}
\alias{spread}
\alias{sym}
\title{gravity exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of gravity.}
\description{
gravity exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of gravity.
}
