% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighstargazer.R
\name{lightstargazer}
\alias{lightstargazer}
\title{Strip regression objects to reduce memory needs}
\usage{
lightstargazer(light_model, ...)
}
\arguments{
\item{object}{A stripped regression by
\link{strip} function. 
Accepted classes are \link[stats]{glm} or
\link[pscl]{zeroinfl}}
}
\value{
A stargazer object
}
\description{
Remove elements that are more related
 to prediction than model to reduce
 memory needs to store a glm output
}
