% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_summary.R
\name{stata_summary}
\alias{stata_summary}
\title{Stata-consistent summary statistics}
\usage{
stata_summary(object, robust = FALSE, ...)
}
\arguments{
\item{robust}{(Type: logical) Determines whether a robust
variance-covariance matrix should be used. By default is set to \code{FALSE}.}

\item{fit}{(Type: lm) Regression object obtained by using the estimation methods from this package 
or a generic method such as \code{lm} or \code{glm}. 
Some particular classes (\code{gpml}, \code{nbpml} and \code{nls}) don't return R squared and F statistic.}
}
\value{
summary \code{lm} object
}
\description{
Summary of estimates written to be consistent with Stata methods.
Adapted after a function written by Isidore Beautrelet
\url{https://raw.githubusercontent.com/IsidoreBeautrelet/economictheoryblog/master/robust_summary.R}
}
\examples{
\dontrun{
data(gravity_no_zeros)

fit <- ols(dependent_variable = "flow", 
 distance = "distw",
 additional_regressors = c("rta", "contig", "comcur"),
 income_origin = "gdp_o", income_destination = "gdp_d", 
 code_origin = "iso_o", code_destination = "iso_d",
 data = gravity_no_zeros)
 
stata_summary(fit, robust = TRUE)
}

}
